ingest/requirements.txt
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.17.0
requests>=2.31.0
tenacity>=8.2.0
